Some arrays are integer

+++ util/flac.py
@@ -43,8 +43,8 @@ class Flac(object):
         # read record number
         tmp = np.fromfile('_contents.0', sep=' ')
         tmp.shape = (-1, 3)
-        self.frames = tmp[:,0]
-        self.steps = tmp[:,1]
+        self.frames = tmp[:,0].astype(int)
+        self.steps = tmp[:,1].astype(int)
         self.time = tmp[:,2]
         self.nrec = len(self.time)
 
@@ -288,8 +288,8 @@ class Flac(object):
 
         suffix = '.%06d.0' % frame
         f2 = open('marker2' + suffix)
-        dead = self._read_data(f2, count=n, dtype=np.int32)
-        tmp = self._read_data(f2, count=n, dtype=np.int32)
+        dead = self._read_data(f2, count=n, dtype=np.uint8)
+        tmp = self._read_data(f2, count=n, dtype=np.uint8)
         phase = self._remove_dead_markers(tmp, dead)
         tmp = self._read_data(f2, count=n, dtype=np.int32)
         ntriag = self._remove_dead_markers(tmp, dead)
